Split compilation
We modified Display_Welcome_Message to be declared in the print.h file
and defined in the print.c file.

version 1.4.0

resolved : #14
